fix found collections but no checkpoint exists
- fixup move test to avoid conflicts from other changes
- send empty Body for collection not found but no checkpoints, and nil if no collection present
